+++ NEWS.md
@@ -1,6 +1,17 @@
+# Version 2.5.22
+## Major Changes
+* Updated AIC GoF to run lognormal and voom for nondiscrete data
+* Added a boxplot of the distribution of aic values to aic_compute
+* Added AIC_median value
+* Updated compute_aic to return a dataframe (rather than multiple) and a boxplox
+
+## Minor Changes
+* Added Harman description to shiny interface
+* Added checkbox for advanced options for Harman
+
 # Version 2.5.21
 ## Major Changes
-* Added harman method in batch correction
+* Added Harman method in batch correction
 
 # Version 2.5.20
 ## Minor Changes